Modify build_embed_structure to return 1 type
There are two ways you can go here, so you could modify the foreach to something like my @field_args_array = $self->build_embed_structure... and then push that onto @tags and store that in an array in your convert_field method and return a reference to that array, or you can go down the route I've suggested here, where the function will return an array_ref that we inflate if we need to push it to @tags. That way convert_field just needs to return the return of build_embed_structure.